+++ app/src/main/java/com/example/partymaker/ui/group/CreateGroupActivity.java
@@ -48,8 +48,8 @@ import com.example.partymaker.utils.auth.AuthHelper;
 import com.example.partymaker.utils.navigation.BottomNavigationHelper;
 import com.example.partymaker.utils.ui.Common;
 import com.example.partymaker.utils.ui.MapUtilities;
-import com.example.partymaker.utils.groupBuilder.GroupBuilder;
-import com.example.partymaker.utils.groupBuilder.GroupDateTime;
+import com.example.partymaker.utils.group.GroupBuilder;
+import com.example.partymaker.utils.group.GroupDateTime;
 import com.google.android.gms.common.api.Status;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationServices;
