+++ app/src/main/java/com/example/partymaker/utils/group/GroupBuilder.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.utils.groupBuilder;
+package com.example.partymaker.utils.group;
 
 import com.example.partymaker.data.model.Group;
 import com.example.partymaker.ui.group.CreateGroupActivity;
